test: unit test stripMarkdown mixin
